+++ extension/CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
+
+## [0.7.1] - 2021-03-17
+
 ### Changed
 
 - Update leparisien integration to support other URL formats